Improve unit tests for Cache\Updater
